requirements.txt
streamlit
inference-sdk
supervision
opencv-python
numpy
Pillow
matplotlib
python-dotenv